UserController: Introduce Bootup Boost
Right after a completed boot a lot of
processes start up.
That can cause slowdowns
which are noticable to the user.

This patch adds the ability to start boosting
for a given duration, with given boost
parameters right before the "BOOT_COMPLETE"
event is dispatched.

The default boost duration is 30 seconds,
which should be a good duration for
most devices.

The boost parameters get transmitted to
QPerformance which forwards it to mp-ctl.

Change-Id: I1154d9a054b89ef9ed056c9db7b472cd815a88b5
Signed-off-by: Alex Naidis <alex.naidis@linux.com>