SystemUI: Internal resource updates
- Generate Pixel navbar for all resolutions, devices follow their
respective resolutions if the folder/drawables exist, causing them to
skip the vectors and fallback on their default configurations. Let's
force this out by removing all instances of the navbar drawables,
leaving the vectors to remain.

- Restore QS/VolumeDialog back to attr/colorPrimary, since we want to
utilize the dark shades of grey merged in a recent commit, there's no
need to allow the newer commits to overwrite that value with it's own,
generating a lighter native Blue Grey color.

Change-Id: I95cf51f91b96669f9efffbfef7ec818cfec8dec7