base: Introduce Anti-Piracy System
The first addition to the new Anti Piracy system. So far when an piracy
related application is installed a notification will be presented to the
user. This notification cannot be swiped away or blocked and will remain
until the piracy related application is uninstalled.

This way we keep things not too instrusive but enough to show the user
that piracy has no place on the software.
The Anti Piracy System only detects LuckyPatcher as of now. More
applications will seek detection in the future.

PS8: Implement overriding fucntion to broadcast our server instead of
the default "Android System" label

Thanks @LorDClockaN for the idea of trolling pirates and for
bringing up the package detection code.

Change-Id: I8ea9280460df06339d2e6c06cf3789acfab8d230