ExtendedRestart: declare availability with OTS
Forces the extended restart menu to be an optional OTS
based feature. Although extended restart is optional as is
with long press only functionality, the first long press will
give regards to the user about what the menu does, instead of
the user "accidently" activating it somehow with no knowledge
of it's functionality.

Change-Id: I77b4533ac69e71f2c16fe509faf3b42065c60508