Doze: Support Hand-Wave Gesture [2/2]
Allows the user the wave their hand across the screen
to trigger Ambient Display. Useful for devices that can't
utilize "Lift to check phone" due to the lack of sensor support.

"Wave to check phone" uses the devices proximity. It also disables
itself when the device is in pocket (in our case, we disable it when
pocket lock is enabled). Thanks to Googles new doze logic, there's no
reason to add this manually for each device anymore.

config_PulseOnHandWave is set to true by default

Todo: Make this use the system's pulse trigger, currently it only
sends a userPulse (Broadcast intent)

Change-Id: If7e41f8b807afa2f4cdac385c207857d2610b9e8