ColorOM: Cleanup accent theme handling
There was a ongoing issue where using setEnabledExclusive
on accents were disabling all theme's including base themes.
That was the default logic on setEnabledExclusive, but created
a bad UX environment in ColorOM, since we separate the two.

So, let's add a statement to only allow setEnabledExclusive
on aoscp accent themes (Themes with isAccent=true in manifest)
and leave the base themes alone.

Also cleaned up the function for accent getters and setters. The code
was too long.

Note: Base themes use setEnabled so this holds no affect on those.

test: works

Change-Id: Ie01f7686979cc1031ac0d4a64fe384a184f249ff