SystemUI: Introduce Ambient Play feature [2/2]
Introducing our alternative to Google's Now Playing feature.
This implementation uses recording logic from Fast Mobile's
"Encore Music" application to pull song fingerprints from a
third party database. Once a match is found with the currently
generated recorded data, the notification and ambient view will
be posted.

The current implementaion requires internet connectivity to contact the
database, but sooner or later, we will alow storing fingerprints in an
offline database as well.

This is based off of @frap129 now playing container backport for the
indication
container and @ezio84 stripped down version of it, which is required.

@Sirhc95: The first iteration was in a separate application with a service,
which registered intents to allow SystemUI to retrieve it's data. This worked
perfectly for notifications, but only worked for the ambient container for about a
day, then stopped working from there on out. I spent days trying to find a resolution
and why it randomly stopped, but couldn't come to a conclusion at this time.

So instead i made this dirty implementaion, which allows it to be
featurized easier, but i hate it. Unfortunately this will have to do for now, until i write a
better way or any other dev can help improve this.

Change-Id: Ifda61b2cd87e09a3873b0c966476a1e9d5369d20