Introduce Ambient Weather [2/2]
A simple weather interface that displays the current weather temp
andcity in Ambient Display and the Lockscreen. This currently uses the
ambient container alongside Ambient Play, but only appears when Ambient
Play results aren't shown.

This uses OmniRom's OmniJaws weather service to fetch weather data. This
is also a recreation of
https://github.com/AICP/frameworks_base/commit/13e9fd05fc77512000dcb5812cfdf0969ae676e6#diff-70fa9685e6a37c48aa825de667646e48

Change-Id: I6cb3205f70e66e5555caa42392caa848c9ae5a64