SystemUI: Remove support for swapping back/recents buttons
Currently causes bootfailure for software key devices and could generate
a swapped layout for hw keys by default with no ability to swap back.
This is due to Tuner API changes in Oreo.

Add this back later only for hw key devices, or both if the NPE can be
resolved

Change-Id: Ib663b06dbf626a5d709cd02de889cd3f5e42aea6