base: Explicitly allow colorized notifications
Previously, in order for a notification to be colorized, it must
be ongoing or in the foreground. Colorized notifications indicate
the high priority and urgency of a notifcation such as incoming calls.

While the "Low Battery" notificaton could be converted to an ongoing
notification to allow this behaivor, the "Optimally Charged" notification
cannot, as it is considered a one time event.

In this case, allow colorizing a notification by simply using "setColorized(true)",
by removing the foreground and ongoing requirements.

Change-Id: I01e9e4feb28a084a4956736d981e03ff1cbe0d39