Network traffic: stop the handler if screen is off
no need to keep the net monitor working if screen is off
and waste system / battery resources

Network traffic: skip more useless handler calls

Move net monitor to expanded statusbar header

Preview: https://i.ibb.co/1TCCwV8/455010636-14778.jpg

Actually, with notch or low dpi and more statusbar icons we have not so
much space there to show the net monitor.
A notch can hide it and the system can now handle only a fixed max number of
notifications and status icons, so when net monitor shows itself it can force
the system to drop some icons to show the infamous dot.

Let's also consider we usually check the monitor to see if the data
transfer is working as expected and the speed is proper, so it's not
an info we want to see always (it's also distracting when reading
something).

So, let's move it to expanded statusbar.
There we have also some room to add line spacing and more info.

Net monitor: fix text color on light theme

Change-Id: I7c10619032bebf5587b4474e15e1fe4c37430615