services: Add support for Huawei fingerprint navigation
Adds support for the fingerprint navigation feature on Huawei
devices. Here, we set the sys.fpnav.enabled prop based on the
state of the status bar. This way, the navigation gestures will
only work when the navigation bar is disabled.

Also set the hw prop as well, since the system reads this for certain
functions. This will also maintain proper compatibility for apps that
read from this prop.

Change-Id: Ic65ae5c7950460f4835b5cb955a89be218b343e7