services: Override qcom hw prop based on navbar state
Set's the hw prop based on the state of the navigation bar to maintain
compatibility for apps, such as Snapchat. Check config_showNavigationBar
to determine to the default state on navbar on first boot. If true, set
the navbar enabled by default.

Based on:

https://github.com/CypherOS/frameworks_base/commit/c1e7fa2f3ccd26bfa12c1f8d0dbca7d7bf359759

Also adds a check for DeviceHWKey support to avoid bootloops on devices
that don't contain this prop nor imports qcom/sepolicy

Change-Id: I8c92373b87d1621a345abe4176c38cddf61a9fd6