Introduce AmbientPlay Music Recognition
Ambient Play is a music recognition service that notifies the user
of songs playing nearby. This is based on Google's now playing but
written completely from scratch. This currently only supports online
recognition with plans to support offline recognition.

Ambient Play is an exclusive feature to CypherOS and can/is only to be
used on this platform. Our MiCode private information is licensed under
GPL and is not to be tampered or replicated.

This commit also adds support for Ambient play history by @jhenrique09

Change-Id: Ie435892426d8cfed4f7527c009e66b0327fd7a0a