Introduce Live Ops
Live Ops is a system that gathers a set of live operations for apps 
and presents them to the user.

Ex. If an app is using your microphone, the user will be presented 
with a microphone icon in the statusbar. Opening the QS will show 
the same icon in the QS header but with the number of apps 
using it. Tapping the chip indicator will open a dialog where the user 
can be properly notified with the name of the application using it
and choose to view the details of the application and it's permissions.

This is very useful for applications that may use sensitive operations 
on the device in the background.

For now, we support 3 ops. Location, Microphone and Camera with 
more to come

- Credits to Lineage since some of the app op getter code 
was referenced from Privacy Guard

Change-Id: I1b0a2b10396f7cc96f72e9c58ef9f85969527191